serdes: hwconfig: Add autonegotiation option for xpcs interfaces
Add "an" property to "xpcs" fields from variable "hwconfig".
Valid options are "an=1" and "an=0". Default (if missing) is 1,
so basically the only logical setting is "an=0" or not specified.

With no "an" or with "an=1" (default), hwconfig fixups will not attempt
to create the "fixed-link" node in the corresponding Linux device tree
node, nor will it delete the "phy-handle" property.

This will allow the corresponding PHY to autonegotiate (if capable)
the link speed with its partner, without need to reboot and reconfigure.

Issue: ALB-10918

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>