serdes: hwconfig: Add 'hwconfig' option 'mode=disabled' for disabling a SerDes module
We are changing the way a SerDes fixup is applied, by adding option
'serdesX:mode=disabled' to 'hwconfig' to disable the corresponding module.
If serdesX configuration is missing from 'hwconfig' then the default
configuration from device tree is applied (and the same is done in case
of invalid serdesX configuration string).

Issue: ALB-10917

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>