u-boot: s32cc gpio driver causes glitch/pulse on enabling output pins
The GPIO driver for the SIUL2 enables the output buffer before setting
up the actual value for the output. This means that a glitch like pulse
is generated on the output until the intended value is set.

This can cause, e.g., unintended resets of peripherals, or, as in the
case of the BlueBox3 a hard power down of the system.

The fix is to first prepare the output registers with the correct
output value before enabling the output buffer.

Issue: ALB-11140

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>

Added issue id to commit description and updated copyright.

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>